+++ CHANGES.md
@@ -6,6 +6,8 @@
   - Checks for trigger_events in already running template
     and does not trigger on events already running
   - Controls stopping of completed spin_ups
+- rt_eqcorrscan.streaming:
+  - Add a lock on `on_data` method and `stream` property for threadsafety.
 
 # 0.0.2
 - First full release. Everything is new!